refactor: consolidate hooks into single hooks/ directory
Move shared hooks from lib/hooks/ to hooks/ — lib/ is conventionally
server-side utilities, not client hooks. Updates 5 import paths.

Closes #369
