+++ apps/console/components/notification-listener.tsx
@@ -2,7 +2,7 @@
 
 import { useCallback } from "react";
 import { toast } from "@/lib/messenger";
-import { useNotificationStream } from "@/lib/hooks/use-notification-stream";
+import { useNotificationStream } from "@/hooks/use-notification-stream";
 import type { BusEvent, BusEventType } from "@/lib/bus/events";
 
 /**
