+++ apps/console/hooks/use-metrics-stream.ts
